requirements.txt
# Core dependencies
mujoco>=3.0.0
gymnasium>=0.29.0
stable-baselines3>=2.2.0
numpy>=1.24.0

# Visualization & logging
matplotlib>=3.7.0
tensorboard>=2.14.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
wandb>=0.16.0

# Configuration
pyyaml>=6.0

# Scientific computing (for terrain generation)
scipy>=1.10.0

# Optional: for Kaggle/Colab video rendering
# opencv-python>=4.8.0
